DisambiguationPatternRule: check for available form/lemma hints before full matching
